TilesetManager - load Tilesets/create Descriptors
for Maps that are of an unconfigured gametype (UFO/TFTD). Therefore such
tileset-configs will be saved in MapTilesets.yml when the MapTree is saved.
They used to get deleted but now it's up to the user to delete them if
he/she doesn't want them. This design bypasses the possibility that user
could deliberately unconfigure a gametype - for whatever reason - and
thereby cause all tileset-configs under that gametype to get accidentally
deleted.

ConfigurationForm - show current resource directories (if configured)
RouteView - show error if attempting to move a node onto another node.